refactor(openomni): re-cut routing on the resolution/execution axis
routing-runtime.ts vs routing-execution.ts was an arbitrary chop:
execution had a single importer (engine) and reached back into
routing-runtime for KernelRouteResolution, while engine.ts itself
hosted ~120 LOC of resolution internals (route stream keying, replay
equivalence, route.decided recording). Move the boundary to the
semantic axis:

- routing-resolution.ts (was routing-runtime.ts): route-state assembly
  + resolveRoute call + execution-plan types, now also owning the
  durable half of resolution moved out of engine.ts (routeStreamId,
  routeDecisionsEquivalent, recordRouteDecided, resolveAndRecordRoute)
  and the IngressRoutingError vocabulary both halves throw.
  resolveKernelRoute becomes module-private (record-before-act means
  nobody may act on an unrecorded resolution).
- routing-execution.ts: effect dispatch only; imports the error and
  the resolution type one-way from routing-resolution.
- engine.ts: pure orchestration (361 -> 240 LOC).

All moved bodies and comments are byte-identical. Test/conformance
imports of IngressRoutingError follow the move.
